+++ compatLib.py
@@ -10,29 +10,6 @@ from time import sleep,time
         timing functions which prefers int(ms) to float(seconds)
 '''
 
-def zip_longest(*args, fillvalue=None):
-    # From:
-    # https://docs.python.org/3/library/itertools.html#itertools.zip_longest
-    # not YET tested in MicroPython..
-    # zip_longest('ABCD', 'xy', fillvalue='-') --> Ax By C- D-
-    iterators = [iter(it) for it in args]
-    num_active = len(iterators)
-    if not num_active:
-        return
-    while True:
-        values = []
-        for i, it in enumerate(iterators):
-            try:
-                value = next(it)
-            except StopIteration:
-                num_active -= 1
-                if not num_active:
-                    return
-                iterators[i] = repeat(fillvalue)
-                value = fillvalue
-            values.append(value)
-        yield tuple(values)
-
 def repeat(object, times=None):
     # repeat(10, 3) --> 10 10 10
     if times is None:
